desktop: split PathOrUrlField into widgets module
Makes possible to reuse this element in bookmarks UI
