+++ desktop/src/gui/open_dialog.rs
@@ -1,7 +1,7 @@
 use crate::custom_event::RuffleEvent;
 use crate::gui::text;
+use crate::gui::widgets::PathOrUrlField;
 use crate::player::PlayerOptions;
-use crate::util::pick_file;
 use egui::{
     Align2, Button, Checkbox, ComboBox, DragValue, Grid, Slider, TextEdit, Ui, Widget, Window,
 };
@@ -9,7 +9,6 @@ use ruffle_core::backend::navigator::{OpenURLMode, SocketMode};
 use ruffle_core::config::Letterbox;
 use ruffle_core::{LoadBehavior, StageAlign, StageScaleMode};
 use ruffle_render::quality::StageQuality;
-use std::path::Path;
 use unic_langid::LanguageIdentifier;
 use url::Url;
 use winit::event_loop::EventLoopProxy;
@@ -515,83 +514,6 @@ impl OpenDialog {
                     keep
                 });
             });
-    }
-}
-
-struct PathOrUrlField {
-    value: String,
-    result: Option<Url>,
-    hint: &'static str,
-}
-
-impl PathOrUrlField {
-    pub fn new(default: Option<Url>, hint: &'static str) -> Self {
-        if let Some(default) = default {
-            if default.scheme() == "file" {
-                if let Ok(path) = default.to_file_path() {
-                    return Self {
-                        value: path.to_string_lossy().to_string(),
-                        result: Some(default),
-                        hint,
-                    };
-                }
-            }
-
-            return Self {
-                value: default.to_string(),
-                result: Some(default),
-                hint,
-            };
-        }
-
-        Self {
-            value: "".to_string(),
-            result: None,
-            hint,
-        }
-    }
-
-    pub fn ui(&mut self, locale: &LanguageIdentifier, ui: &mut Ui) -> &mut Self {
-        ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-            if ui.button(text(locale, "browse")).clicked() {
-                let dir = self
-                    .result
-                    .as_ref()
-                    .filter(|url| url.scheme() == "file")
-                    .and_then(|url| url.to_file_path().ok())
-                    .map(|mut path| {
-                        path.pop();
-                        path
-                    });
-
-                if let Some(path) = pick_file(true, dir) {
-                    self.value = path.to_string_lossy().to_string();
-                }
-            }
-            ui.add_sized(
-                ui.available_size(),
-                TextEdit::singleline(&mut self.value)
-                    .hint_text(self.hint)
-                    .text_color_opt(if self.result.is_none() {
-                        Some(ui.style().visuals.error_fg_color)
-                    } else {
-                        None
-                    }),
-            );
-        });
-
-        let path = Path::new(&self.value);
-        self.result = if path.is_file() {
-            Url::from_file_path(path).ok()
-        } else {
-            Url::parse(&self.value).ok()
-        };
-
-        self
-    }
-
-    pub fn value(&self) -> Option<&Url> {
-        self.result.as_ref()
     }
 }
